refactor(index): move stuff around to make room for non-zero mod.
